dynparquet: Fix buffer serialization
This appears to be a bug in parquet-go that I haven't been able to track
down yet, but for some reason copying appears to be faulty in certain
situations when copying an in-memory buffer to a persistent buffer.
However, when writing row by row, it works. This is likely a bug in the
optimization paths CopyRows attempts to take.