fix query schema not hard-coded
fixes issue in query/engine.go where the schema passed to the physical plan builder was hard-coded to the sample schema.
It now gets the schema from the logical plan.
A new method was added to the logcal plan to return the schema which traverses the tree to get the schema from the TableProvider on either TableScan or SchemaScan steps.
A new method was also added to the TableReader interface which returns the schema