*: Optimize distinct queries with single value per rowgroup
Gathering distinct values of (non optional) dictionary encoded columns
is a constant operation per rowgroup when the dictionary only has a
single value. Additionally distinct projections using binary expressions
such as `timestamp > 0` can be determined without loading the row if the
column index has sufficient data. For example when the index notes that
all timestamps are between 5 and 8, then all values for `timestamp > 0`
are true, and therefore true for any distinct values in the rowgroup.