query/physicalplan: update TestDistinct to check expected rows
This test modification uncovered a bug in the distinct optimized path. NULLs
are not represented in column indexes or dictionaries, resulting in missed
values. This commit falls back to the unoptimized path in the presence of
NULLs in the page. This behavior is temporary until we verify upstream whether
it is expected that NULLs are not represented in the column index or
dictionaries.