query/physicalplan: Use concurrency in the physicalplan  (#202)
* Write rowGroups into channel and enable concurrent TableScan and SchemaScan

* query/physicalplan: Create a Synchronizer using errgroup and channels

* query/physicalplan: Create concurrent Distinct with Synchronizer and final Distinct

* Add a synchronizer with test

This synchronizer now uses an uint64 atomic to count down the number of finished plans.
Only the last plan to call to Finish will then propagate the Finish to the single next plan.

* Aggregation only renames alias column if final stage

Only the last aggregation should rename the value column to e.g. value_sum.
The previous aggregations need to still return with the name value, so that the last stage can find that column by its original name.

* Multi-stage count aggregation needs to be summed up

If we have 3 aggregators that concurrently count some column, and then forward the result. The final aggregation stage needs to sum up the results of the previous stage count.

* Create concurrent plan for distinct, filter and aggregation

Based on a hardcoded concurrency we create multiple concurrent operators that in the end synchronize to call the last callback sequentially.

* query/physicalplan: Create a Synchronizer

* query/physicalplan: Remove obsolete mutex on distinct operator
query/physicalplan: Create multiple concurrent buffers to aggregate from

* query/physicalplan: Update tests to check for expected and seen rows

* query/physicalplan: Run Finish calls in concurrent errgroup

* Revert to old Distinct after rebase

* Fix table.go after rebase

* query/physicalplan: Synchronizer returns error if Finished too often

* Table iterator uses channels and parquet converter

* go mod tidy

* query/physicalplan: Use atomic/sync in synchronize.go

* Don't call callback if record has no rows

* query/physicalplan: Fix TestDistinctPartialScanOptimization

We receive multiple callbacks now and need to check at the end if the correct number of rows have been seen.

* query/physicalplan: Add waitgroup to wait for serializing row groups

* query/physicalplan: Actually wait for the waitgroup