logictest: add datadriven logic testing framework
This commit adds a simple datadriven logic testing framework with the ability
to create tables, insert data, and query data. Some existing distinct and
aggregation tests have been copied and modified to use the new testing
framework.

The advantage of a framework like this is that it makes tests easier to
understand and write, simplifying bug resolution and lowering the bar to
writing tests.