sqlparser/logictest: add datadriven testing (#211)
* sqlparse: add experimental SQL parsing

This commit adds a sqlparse package with the initial functionality to parse
SQL into FrostDB logical plans. Distinct, Aggregate and Filters are all
supported. SQL is parsed using TiDB's standalone MySQL parser. Given the
simplicity of the SQL we support, there are no variant-specific corners we're
painting ourselves into yet.

* logictest: add datadriven logic testing framework

This commit adds a simple datadriven logic testing framework with the ability
to create tables, insert data, and query data. Some existing distinct and
aggregation tests have been copied and modified to use the new testing
framework.

The advantage of a framework like this is that it makes tests easier to
understand and write, simplifying bug resolution and lowering the bar to
writing tests.