frostdb: miscellaneous concurrency fixes
This commit fixes a couple of issues introduced with concurrency:
- Adds a case to select on context cancellation when collecting row groups to
respond to cancellation while blocked on sending a row group.
- Adds a missing loop to read rowGroups in SchemaIterator. Previously, only
the first row group produced would be read, then the reader goroutine would
exit, resulting in the writer goroutine blocking on sending over a channel.
- Closes the row group channel before waiting on the reader WaitGroup. This is
a signal that the producer has finished producing all row groups. Previously,
this could result in a deadlock where the producer goroutine is waiting for
the reader goroutines to finish, but these are waiting for the producer to
close the channel.
- Puts row group collection into an error group. This allows for aborting row
group collection if any reader goroutine encounters an error (and vice-versa).