pqarrow/builder: optimize Parquet to Arrow conversion (#218)
* pqarrow: use builder.RecordBuilder instead of array.RecordBuilder

This gives us the freedom to modify the RecordBuilder layer, a useful
abstraction for Parquet -> Arrow conversion.

This commit also removes ParquetRowGroupToArrowRecord, since it was only used
in tests and is now superseded by using a ParquetConverter.

* builder: add OptBinaryBuilder

This builder accepts appending flat binary + offsets data and constructs an
arrow array when NewArray is called, just like an array.Builder. This allows
us to bypass constructing [][]byte slices when reading from parquet and copying
data back to a flat bytes buffer when writing to a BinaryBuilder.

* builder: add OptInt64Builder

This commit adds and uses an OptInt64Builder when scanning int64 values. The
commit also adds a ListBuilder that wraps an array.ListBuilder in order to
use optimized builders as value builders for list types.

* pqarrow: check whether page has any nulls instead of null definition

Columns with optional values may have pages that have no values. This commit
checks the number of nulls in a page instead of the column definition in order
to use the non-null scan fast path when possible.