*: add automatic leak detection to the frostdb package (#222)
* *: upgrade parquet to 771b3e

* *: use parquet's new debug tag to disable finalizers

This allows us to add leak verification in our tests to detect goroutine leaks
earlier in the development cycle. There have been cases where these goroutines
leaked and held on to memory, preventing the GC from collecting it and causing
an OOM kill.

* frostdb: add TxPool.Stop to signal cleaner goroutine to stop

This was a leaking gorouine that was never stopped. Cleaning up these internal
goroutine leaks will allow us to run leak checks during tests to avoid other
goroutine leaks.

* *: add missing ColumnStore.Close calls

* frostdb: add goroutine leak detector in TestMain

This commit also fixes a bunch of leaks in tests so that the leak detector can
work.