frostdb: add simple leveled compaction to parts
This commit adds a compaction level to parts. This improves memory usage by
ignoring parts that were already compacted previously and already have
optimally-sized row groups.

The other big change in this commit is making the split code a method on a
Part rather than a Granule, since the latter assumed it was composed of only
one part. Additionally, a bug was fixed in the split code where the desired
row group size was ignored, thus writing a single row group to the split
results.