deps: update parquet-go to db0f5ce (#272)
* deps: update parquet-go to db0f5ce

This pulls in synchronous page reading as well as a bunch of performance
improvements to MergeRowGroups (useful for compaction).

There was a breaking change in the update where the SortFuncOf API was removed,
resulting in a rewrite of the dynparquet.Schema.Cmp method to use a
parquet.Schema.Comparator.

* frostdb: read object storage buckets using the async option

A recent change in parquet changed the default page reading mode from
asynchronous to synchronous. When reading object storage, we explicitly want
to read pages asynchronously given the high cost of reading.