pqarrow: add ArrayConcatenator
This is a simple helper struct to use in the place of a builder when the caller
just wants to accumulate arrays and concatenate them at some point in the
future. The need for this was discovered while building the OrderedAggregate.
Calling NewArray on the builder and storing the result is not memory safe since
following uses of the builder would overwrite the previous result.