physicalplan: add OrderedSynchronizer (#301)
* physicalplan: add leak testing to physicalplan package

* physicalplan: add OrderedSynchronizer

This commit adds an OrderedSynchronizer component that merges ordered records
pushed from many concurrent inputs and pushes the merged record to a single
output.

The strategy this commit implements is to block the inputs until all inputs
have pushed a record. This is the most straightforward strategy since the other
option is to store the pushed records, but that requires copying all data which
might be more expensive.