physicalplan: add partial ordering support do the OrderedAggregate  (#299)
* pqarrow: move getGroupsAndOrderedSetRanges from OrderedAggregator to arrowutils

This will need to be reused when getting groups for merged records and it made
more sense to generalize + move the comparison logic to a single package.

* pqarrow: add ArrayConcatenator

This is a simple helper struct to use in the place of a builder when the caller
just wants to accumulate arrays and concatenate them at some point in the
future. The need for this was discovered while building the OrderedAggregate.
Calling NewArray on the builder and storing the result is not memory safe since
following uses of the builder would overwrite the previous result.

* query: extract final stage aware aggregation into a helper function

* physicalplan: add partial ordering support do the OrderedAggregate

This commit adds the missing partial ordering support to the OrderedAggregate.
Instead of returning a record on each call to Callback as was previously done,
the OrderedAggregate now buffers the aggregation results and groups for each
ordered set and then merges the results on Finish.