physicalplan: plan ordered aggregations (#307)
* physicalplan: support final stage aggregations in OrderedAggregate

* physicalplan: fix OrderedSynchronizer

The OrderedSynchronizer was broken in cases where inputs would call Finish
without calling Callback. This commit fixes that and updates the test to call
Finish.

* physicalplan: plan OrderedAggregate

This commit plans the OrderedAggregate in cases where the group columns are a
prefix of the ordering columns.

* physicalplan: use expressions instead of indices in OrderedSynchronizer

This commit fixes a bug in ordered synchronization where input streams were
merged incorrectly given they could have different schemas because of dynamic
group by columns. This commit adds an ensureSameSchema step to records before
merging them so that all dynamic columns are accounted for before merging.

* physicalplan: put OrderedAggregate planning behind an option

Given this is a new execution component, it seems reasonable to offer ordered
aggregations as an option that is disabled by default and eventually move to
planning ordered aggregations by default.