wal: add batch to log message when encountering write error (#308)
tidwall/wal can refuse to write a batch due to out of order indexes, but it's
hard to tell the underlying cause just from the error message. Printing the
batch that we attempted to write will help diagnosing the cause. The log
message will be very verbose, but since it is only logged on error, that should
be fine.