wal: fix possible deadlock
This commit fixes a deadlock when a WAL client would attempt to log an entry
with a txn id lower than the next expected txn. This commit fixes this deadlock
by simply dropping the minimum entry in the queue if it is lower than the next
expected txn. This could lead to missing data so a warning is logged, but the
result is better than never making progress on the WAL and losing all data (not
just a single entry) after a txn id due to preemption.