wal: miscellaneous fixes (#323)
* wal: fix "unsupported value type" in logging message

* wal: fix possible deadlock

This commit fixes a deadlock when a WAL client would attempt to log an entry
with a txn id lower than the next expected txn. This commit fixes this deadlock
by simply dropping the minimum entry in the queue if it is lower than the next
expected txn. This could lead to missing data so a warning is logged, but the
result is better than never making progress on the WAL and losing all data (not
just a single entry) after a txn id due to preemption.

* wal: make run method private

This was previously public, allowing a caller to call Run. However Run would
already be called on Open, causing two readers to read from the same queue.
Since we don't need to allow callers to explicitly call Run, it's safer to make
this method private.