frostdb: skip replaying WAL writes for tables that have already been persisted (#340)
When replaying writes, we would only check if the table did not exist to skip
replaying a write, implying that the new table block entry had been truncated
and therefore persisted. However, sometimes the database can only truncate
entries if ALL tables have been persisted and only then up to the minimum
transaction persisted across all tables. This means that even if a block
belonging to a larger table is persisted, writes belonging to this block might
not be truncated from the WAL yet. This commit skips these writes if it sees
that the block the writes belong to were already persisted, reducing
unnecessary memory pressure on the databse.