*: add snapshots
This commit adds the ability to write a database snapshot at a given tx to disk
and load a snapshot into a database. This commit does not yet add a
public-facing API for snapshots.

The motivation for this change is to improve use cases where users enable the
WAL. Snapshots aim to solve two problems:
1) WAL replay can be very slow given that each insert needs to be replayed to
get to a "final" state. Snapshots can be loaded into a database in O(1) time
with respect to the number of inserts performed against the database instead
of the O(n) inserts required during a WAL replay.
2) With multiple tables, the WAL can only be truncated when all tables have
been persisted/rotated out and only at the minimum persisted tx across all
tables. If the user has a small table that doesn't change often and is
therefore not persisted/rotated out, the WAL can never be truncated. Snapshots
allow us to persist these small tables and truncate the WAL at the snapshot
txn.