Expect io.EOF and at the end of files
We expect an io.EOF at the end of files  and in that case return total bytes read. This is not an issue for the header of ranges within the file, however, especially when reading the parquet footer we read until the end of the file and this would previously return io.EOF even though it's expected.