frostdb: enable snapshots (#371)
* frostdb: fix txpool race on shutdown

Close would close a channel that is written to from another goroutine, which
is a race condition since a write could panic. This commit changes the txpool
to use a context with a cancel function to shut down instead.

* frostdb: enable snapshots

This commit adds the option to enable snapshotting a database whenever a block
exceeds a given size or is rotated out.
A snapshot is taken of the db state and stored in the snapshots/ folder. The
use of the WAL is not modified yet, even though it could be truncated. This
will be done in a future change once snapshots have been running for a while.

Note that old snapshots are not cleaned up yet, which gives us more txns to
reload at in case of issues. In the future, a cleanup policy will be
implemented.