*: miscellaneous recovery fixes (#393)
* wal: remove nextTx and txmtx

These were variables the WAL was using to track the next index to log. However,
the underlying WAL library already has this information.

* *: improve logging around db recovery

This logging was previously pretty verbose. Although it is important to keep
the log level at info, some messages were downgraded to debug and others were
merged to improve the experience.

* *: set table WAL to db WAL on recovery

Tables recovered from a snapshot are created with a NoopWAL by default. On
recovery, these pointers need to be updated to the DB wal, otherwise future
writes will not be logged to the WAL.

* *: add TestReplayBackwardsCompatibility