*: recovery fixes (#400)
* db: fix replay backwards compatibility test structure

* *: make snapshotting db on block rotation optional

There were cases where this wouldn't make sense. On shutdown, for example, the
database was snapshot n times where n is the number of tables in sequential
transactions. This greatly increases shutdown times without any benefit.

* *: overwrite table WALs after opening WAL

This was previously done only after loading a snapshot, but a DB configured
with object storage could create tables with a NopWAL as well.