bench: add BenchmarkQuery/Filter (#423)
* workflows: update go version to 1.20

* bench: add BenchmarkQuery/Filter

This sub-benchmark runs within the BenchmarkQuery setup to measure how fast a
simple filter runs. B/msec are also reported, which is something that we're
tracking in our OKRs (i.e. scan layer throughput). These bytes are measured as
total block size, so if we discard scanning granules, that counts as processing
that granule which is nice because it captures how well we're doing in scan
layer filtering before parquet conversion.