db: move snapshots and wal directories to trash (#426)
* wal: fix metrics typo

* db: move snapshots and wal directories to trash

Previously, db.Close would call os.RemoveAll on the snapshots and wal
directories. This is an O(n) operation, which could lead to some leftover
files if we exceed the graceful shutdown period.

This commit changes the removal to move both directories to a trash/ directory,
which is an O(1) operation. This trash/ directory can be removed safely on
startup if it still exists.

Additionally, this commit moves closeInternal to before removing the
directories. This ensures that the WAL is flushed and closed before deletion.