wal: perform truncation in run (#436)
* wal: move queue into protected struct

This is an idiomatic way to make it more obvious that the queue cannot be
accessed without locking the mutex first. Some more protected data will be
added in a future commit. This commit also removes unnecessary pointers.

* wal: perform truncation in run

Previously, we would synchronously truncate even though we asynchronously wrote
to the WAL. This caused truncations to fail with high likelihood given that the
transaction to truncate at was likely not yet logged.