*: reclaim disk space on successful snapshot (#437)
* *: refactor snapshot file iteration

This extracts snapshot reverse file iteration into a helper method that
executes a callback against each snapshot file.

* *: reclaim disk space on successful snapshot

This commit reclaims disk space after performing a successful snapshot on
block rotation and insert. The code is extra-careful to open a snapshot file
and check the checksum before using that transaction to truncate old entries
from the WAL and old snapshots.