wal: fix race on truncation (#439)
A previous commit moved truncation to the run loop. However, the code would
only perform a truncation if a new record with the truncate txn was popped off
the heap. There could be cases where the record to truncate at has already
been written to the WAL, which is handled by this commit.