dynparquet: reduce allocations (#434)
* dynparquet: remove allocation on call to SortingColumn.Path

This showed up in production profiles as 5% of cpu usage. A sorting column's
path can be computed on instantiation once and stored, since it is immutable.

* dynparquet: pool parquetSortingSchema

This showed up as a hot path (a big part of ~70% CPU time in insertions) in
production profiles. This change improves BenchmarkInsertSimple by 50%.

* dynparquet: pool DynamicParquetSchema

This should reduce allocations on a hot path.