*: add option to manually rotate table blocks (#474)
Previously, table blocks were only rotated when reaching a certain size. This
commit allows a user to disable this behavior in order to manually trigger a
block rotation via RotateBlock.

A skipPersist option has also been added to RotateBlock so that blocks are not
persisted on rotation even if the user specified bucket storage.