db: apply options when re-opening database (#485)
When opening a database with an on-disk WAL, the first instantiation is a
side-effect of an internal `DB` call in order to recover from wal/snapshots. It
doesn't make sense to pass user options through this frostdb.New call since the
user needs to specify per-db options.

Instead, this commit applies options to a db even if the db already exists,
which handles cases where an internal db creation had no options, but the first
user-triggered DB call does.