wal: allow Truncating past end of WAL
This commit implements the logic to allow truncating past the end of the WAL,
i.e. resetting the WAL entirely. The only condition is that the next record
logged MUST be the txn immediately following the txn at which the WAL was
truncated.

This is very useful for recovery where we have a snapshot with a txn k and the
last index of a WAL is j < k-1, so we don't really care about the WAL entries
anymore.