wal: allow Truncating past end of WAL (#487)
* deps: update polarsignals/wal

This pulls in a commit that allows calls to TruncatFront past the last index
in order to reset the WAL.

* wal: allow Truncating past end of WAL

This commit implements the logic to allow truncating past the end of the WAL,
i.e. resetting the WAL entirely. The only condition is that the next record
logged MUST be the txn immediately following the txn at which the WAL was
truncated.

This is very useful for recovery where we have a snapshot with a txn k and the
last index of a WAL is j < k-1, so we don't really care about the WAL entries
anymore.

* *: rework recovery to reset the WAL when a snapshot is loaded

Currently, there were cases where we had a snapshot with a txn k and a last
index of a WAL j < k-1, which resulted in an unusable WAL since we would resume
from the snapshot txn without the ability to log to the WAL. Now that we can
Truncate past the end of the WAL, the code is reworked to implicitly reset the
WAL for correct logging if we have a more up-to-date snapshot.