*: reset the WAL to zero in resetToTxn
Previously, we would use the given txn to truncate the WAL. However, this txn
could be less than the last index, causing no records to be logged until the
high watermark caught up with the last index.

This commit performs an unconditional truncation using the max uint64,
resulting in a full reset of the WAL. This means we cannot rely on resetToTxn
to truncate the WAL in recover, so that code has been changed to perform
manual WAL truncation.