*: handle column chunks with contained nulls in BinaryScalarOperation
Previously, nulls were only handled for the logicalplan.OpEq operator. There
were cases where the presence of nulls in the application of other operators
(e.g. logicalplan.OpGt) would lead to an unsupported value panic in the
comparison function. Indeed, null comparisons are undefined according to SQL
semantics so the safest bet right now is to simple have null values pass the
check and delegate the filtering to the execution engine.

There is probably some optimizations we can do here once we define the
semantics we want, but this commit opts for safety first as a first step.