*: filter out column chunks with some null values
Previously, if a column chunk had at least one null value, we would
automatically further filter it in the execution engine. However, there are
cases where we can rely on min/max values as long as they are non-null and the
column chunk is not all nulls to filter out column chunks. This greatly reduces
load on the execution engine.