db: start compactor pool before replaying WAL
Previously, the compactor pool was not running during WAL replay. This could
cause memory blow-ups on the first compactions after a WAL replay since there
would be a lot of l0 parts processed at the same time.

Starting the compactor pool before WAL replay causes WAL replay to slow down
a bit but with the good property that data is compacted during WAL replay and
memory spikes are avoided. Also, recovery code already expected the compactor
pool to be started.

This commit also removes the CompactAfterRecovery option given it is not needed
any more.