db: clean up snapshots greater than valid snapshot
Since we added snapshots on close, it is possible that a large enough block
size causes every close to produce a large invalid snapshot. This can cause
an instance to run out of disk space quickly if not handled.

This commit simply removes all snapshots greater than the loaded snapshot since
they are invalid by definition.