lsm: hold read lock on iterate (#557)
* lsm: hold read lock on Iterate

Otherwise, snapshots will not hold a read lock over the index when iterating.

* *: handle 0-size object storage blocks

There are sometimes errors when serializing a block. This would cause us to
create a file but not delete it. This commit deletes an object storage file on
serialization failure.

Additionally, this commit ignores empty object storage files on scans.