wal: size batches according to segment size and track WAL queue
Previously, we weren't controlling the total size of WAL entries written at
once to the WAL. The underlying WAL does not rotate a file until after a batch
is written, so could allocate a lot (we've seen >GB in production) for the
commit buffer.

This commit should alleviate in-use bytes for cases where lots of [large]
entries are written.