wal: don't wait for ticker to fire to clear pending txns
Previously, if we found a txn below the nextTx, we would pop it but continue
waiting for the timer to fire again before popping the next one. This is not
necessary and can lead to high memory usage.

This commit also avoids pushing these requests to the queue in the first place.