db: perform a snapshot if WAL contains persisted entries (#569)
* db: set uncompressed insert size after recovery

Previously, the uncompressed insert size would not be initialized correctly
after recovery, this could cause L0 size to grow

* db: perform a snapshot if WAL contains persisted entries

When rotating or persisting a block, a TableBlockPersisted entry is logged in
the WAL and an async snapshot is performed. If on recovery, the entry still
exists in the WAL, it means that a snapshot was not performed, since it would
have truncated the WAL. This commit ensures the WAL is cleaned up in these
cases, since we have seen that this could lead the WAL to grow indefinitely.