Remove (*Table)Write and introduce (*GenericTable[T]).Write
This commit removes `(*Table).Write` in favor of `(*GenericTable[T]).Write`

`GenericTable` offers a lot of improvements from the previous API. One of the
biggest advantage is there is no longer need to define schema before hand
instead the schema is part of the struct that is written in the form of struct
tags.

Example code has been refactored to use `map[string]T` for dynamic columns,
however slice struct dynamic columns still works, this is intentional to avoid
the PR from blowing out of proportion.

`GenericTable` embeds `Table` so from user perspective only few changes are
needed to work with the new API.