LSM recovery
After creating a new fileCompaction object for an LSM level, recover any
parts that are found in that file. Insert those parts into the
corresponding level in the index.

Parts that are older than a committed part in a higher level are dropped
as they must have been compacted into that level.