Return sorting metadata on (*records.Build[T]).NewRecord
This commit wraps returned `arrow.Record` with sorting column information. We
already have sorting implementation for `arrow.Record`.

The reason we need this is to eventually allow `GenericTable` to automatically
build and sort `arrow.Record` before inserting them to the `lsm`, removing the
requirement that values needs to be sorted before being written.