Persistent lsm (#667)
* index/lsm: Order inserts

Makes the LSM inserts ordered by tx.
During compaciton it will use the provided wait function
to ensure that the watermark is up to the highest tx in the compaction
list thus ensuring that compactions are of contiguous writes.

* LSM recovery

After creating a new fileCompaction object for an LSM level, recover any
parts that are found in that file. Insert those parts into the
corresponding level in the index.

Parts that are older than a committed part in a higher level are dropped
as they must have been compacted into that level.