move memoryBlocks call outside collect row groups
Previously we could buffer the row groups we were reading but would
decrement the pending readers wait group.

This doesn't play well with dropPendingBlocks with file based indexes.
Because we can close the files cause we think readers of the block are
done, but we still need to read those row groups.

Moving it outside of the collect function means we know that there are
no more readers for certain.