Add parca schema sample
This commit adds a `Parca` struct sample defining the current `parca` database
schema.

This is for use in tests, as well as giving perspective of production outlook
for some features. For instance we mostly focus on `*array.String` or
`*array.Binary` dictionary, but we see `*array.Int64` dictionary in this schema.

The plan is to eventually remove `parquet` -> `arrow` conversion during
ingestion without breaking backward compatibility, this seems like a good start.