Fix index deadlocks (#711)
* snapshot: Don't acquire write TX

There's a deadlock with snapshots and compactions with the new index.
We don't need the snapshot to have a write tx, but instead a read tx.

Also remove the WAL log for snapshots since we don't utilize them at
all.

* Fix: snapshot deadlock; part release

- Snapshot deadlock
    much like the previous commit we need to use a read tx for the
snapshot and not a write tx. In addition after the WAL is logged to with
the block rotation we need to commit the tx so as to not deadlock while
snapshotting

- Part release
    If a read gets cancelled or panics we could strand parts in the
channel. Add a defer'd function to always drain the channel and release
parts there.

* index: Don't lock longer than needed

In addition reset the size gauge when creating a new LSM

* unit test fix

Fixes a snapshot test that relied on the snapshot increasing the tx

* remove stale comment