query: add global aggregation function pushdown optimization for Max
When reading Particulates from blocks (either in-memory or object storage), we
can drastically reduce the runtime of global aggregations with no filters. This
commit implements the optimization pushdown as a filter.

In case of Max (implemented in this commit) for example, the TrueNegativeFilter
memoizes the maximum value seen and filters out any Particulates that will
definitely not contribute to the aggregation function (i.e. whose max value is
less than the max value seen so far). This drastically reduces global Max
queries. For example, on 1.5GB of data, a global Max on an unsorted int64
column went from 900ms to 50ms.