Fix: double snapshot data loss
Previously if we took a snapshot, stopped writing and shutdown the
database we would take another snapshot. This snapshot would fail
which would cause it to remove the valid snapshot directory.

This changes the behavior of snapshots to first check if a snapshot
already exists. If it does validate it and if the snapshot is valid
abort taking a new one. If the snapshot is invalid then remove it and
perform the snapshot.