query/expr: Infer type from arrow.Scalar for parquet Value
When comparing we want to use arrow's scalar.Scalar as the type supports a lot more underlying types, like unsigned integers.

Therefore, we want to convert parquet.Values to arrow scalar.Scalar and then compare. That way we can ensure that the unsigned integers, for example, are properly treated and don't overflow.