query: Add option to ignore missing columns.
Missing columns are typically treated as a null column. However when
filtering on a Iceberg's manifest partition missing columns cannot be
treated as null because the partition only containes the columns that
are configured in the partition spec.

So filtering manifests we just treat missing columns as a positive hit
since we can't make a negative decision based on it not being included.