db: ignore writes below new block transaction on recovery
Previously, the recovery code was ignoring writes below the persist transaction
of a table block. However, this could lead to dropped writes since the active
block is swapped atomically on rotation *before* the old block is persisted.
Writes in between these two events would be written to the new table block, but
ignored on recovery given the recovery code assumed they were in the old table
block.