dst: add deterministic simulation test
Running this test by itself is just a normal integration test. To run this test
deterministically use POLARSIGNALS_RANDOM_SEED, the WASM build target, and
github.com/asubiotto/go. For now, only the test is checked in to reduce the
change surface area and is skipped if POLARSIGNALS_RANDOM_SEED is not
specified.