dst: add deterministic simulation test (#865)
* deps: update polarsignals/wal to 1cd4b81c9b88

This allows building polarsignals/wal with WASM.

* snapshot: add Snapshot public method to DB

This allows callers to trigger a snapshot that is written to the snapshots
directory. Taking snapshots was previously only exposed through a
WriteSnapshot method that required a writer. This commit makes taking snapshots
easy in tests.

* table: sleep one millisecond when generating ULID

ULID times have millisecond precision. To avoid executing multiple loop
iterations unnecessarily, this commit explicitly sleeps a millisecond.

This is also required for DST tests where time is faked and only advanced if
all goroutines are sleeping.

* *: add WithTestingOptions option to ColumnStore

This commit organizes ColumnStore option for exclusive use in tests. It also
adds WithTestingWalWrapper so that tests can provide a WAL implementation that
wraps the real deal.

* wal: refactor processing code into process() function

This is both cleaner and allows the WAL to proactively process pending entries
on close.

Additionally, it plays better with deterministic simulation tests because the
run goroutine spins indefinitely when closing, which doesn't allow faketime to
advance.

* dst: add deterministic simulation test

Running this test by itself is just a normal integration test. To run this test
deterministically use POLARSIGNALS_RANDOM_SEED, the WASM build target, and
github.com/asubiotto/go. For now, only the test is checked in to reduce the
change surface area and is skipped if POLARSIGNALS_RANDOM_SEED is not
specified.

* wal: add WithTestingLoopTicker option

This allows callers to control when the WAL main loop ticker fires, either
delaying or forcing processing. This is very helpful for DST tests to trigger
randomly batched writes at will.