db: ignore writes below new block transaction on recovery (#864)
* db: ignore writes below new block transaction on recovery

Previously, the recovery code was ignoring writes below the persist transaction
of a table block. However, this could lead to dropped writes since the active
block is swapped atomically on rotation *before* the old block is persisted.
Writes in between these two events would be written to the new table block, but
ignored on recovery given the recovery code assumed they were in the old table
block.

* db: fix WAL off-by-one truncation error

Previously snapshots were performed using a write txn and the WAL was truncated
so that that txn would be the first txn in the truncated WAL. However,
snapshots were changed to use a read txn so if a write at txn k was included
in the snapshot, a truncated WAL would still contain this write as the first
write in the WAL, resulting in duplicate data after recovery.